+++ Changes.md
@@ -5,6 +5,7 @@ Version 3.3.1.3.5
   2. When user in the default section, prevent all sections from loading when editing is turned on - https://www.screencast.com/t/iCX4cuKc0M0.
   3. Always show 'Jump to' menu on a single section page.
   4. Move icon and always show Jump To menu when view single section in edit mode.
+  5. Jump to menu on multiple section page.
 
 Version 3.3.1.3.4
   1. Stage 8.